requirements.txt
pytest==7.1.2
allure-pytest==2.9.45
selene>=2.0.0rc2
selenium~=4.25.0